Refactor: 흩어져 있던 hook들을 hooks 디렉토리로 위치변경
